+++ mgmt/tests/reconcile.rs
@@ -1,11 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Open Network Fabric Authors
 
-#![cfg(feature = "bolero")]
-
 use dataplane_mgmt as mgmt;
 
-use caps::{CapSet, Capability};
+use caps::Capability;
 use fixin::wrap;
 use interface_manager::interface::{
     BridgePropertiesSpec, InterfaceAssociationSpec, InterfacePropertiesSpec, InterfaceSpecBuilder,
@@ -13,131 +11,18 @@ use interface_manager::interface::{
     MultiIndexInterfaceSpecMap, MultiIndexVrfPropertiesSpecMap, MultiIndexVtepPropertiesSpecMap,
     VrfPropertiesSpec, VtepPropertiesSpec,
 };
-use interface_manager::netns::swap_thread_to_netns;
 use mgmt::vpc_manager::{RequiredInformationBase, RequiredInformationBaseBuilder, VpcManager};
 use net::eth::ethtype::EthType;
 use net::interface::AdminState;
 use net::vxlan::Vxlan;
 use rekon::{Observe, Reconcile};
-use rtnetlink::NetworkNamespace;
 use rtnetlink::sys::AsyncSocket;
 use std::net::Ipv4Addr;
-use std::panic::{RefUnwindSafe, UnwindSafe, catch_unwind};
 use std::sync::Arc;
 use std::time::Duration;
-use tracing::{error, info};
+use test_utils::{in_scoped_netns, with_caps};
+use tracing::info;
 use tracing_test::traced_test;
-
-/// Fixture which runs the test in a network namespace of the given name.
-fn run_in_netns<F: UnwindSafe + Send + FnOnce() -> T, T>(
-    netns_name: impl AsRef<str>,
-) -> impl FnOnce(F) -> T
-where
-    T: Send,
-{
-    move |f: F| {
-        let netns_path = format!("/run/netns/{netns_name}", netns_name = netns_name.as_ref());
-        std::thread::scope(|scope| {
-            std::thread::Builder::new()
-                .name(netns_name.as_ref().to_string())
-                .spawn_scoped(scope, || {
-                    with_caps([Capability::CAP_SYS_ADMIN])(|| unsafe {
-                        swap_thread_to_netns(&netns_path)
-                    })
-                    .unwrap_or_else(|e| panic!("{e}"));
-                    catch_unwind(f).unwrap()
-                })
-                .unwrap()
-                .join()
-                .unwrap()
-        })
-    }
-}
-
-/// Fixture which creates and cleans up a network namespace with the given name.
-fn with_scoped_netns<F: 'static + Send + RefUnwindSafe + UnwindSafe + Send + FnOnce() -> T, T>(
-    netns_name: impl 'static + Send + UnwindSafe + RefUnwindSafe + AsRef<str>,
-) -> impl FnOnce(F) -> T
-where
-    T: Send,
-{
-    move |f: F| {
-        let runtime = tokio::runtime::Builder::new_current_thread()
-            .enable_io()
-            .enable_time()
-            .build()
-            .unwrap();
-        with_caps([Capability::CAP_SYS_ADMIN])(|| {
-            runtime.block_on(async {
-                let netns_name_copy = netns_name.as_ref().to_string();
-                let Ok((connection, _, _)) = rtnetlink::new_connection() else {
-                    panic!("failed to create connection");
-                };
-                tokio::spawn(connection);
-                match NetworkNamespace::add(netns_name_copy).await {
-                    Ok(()) => {}
-                    Err(err) => {
-                        let netns_name = netns_name.as_ref();
-                        panic!("failed to create network namespace {netns_name}: {err}");
-                    }
-                }
-            });
-        });
-        let ret = catch_unwind(f);
-        with_caps([Capability::CAP_SYS_ADMIN])(|| {
-            runtime.block_on(async {
-                match NetworkNamespace::del(netns_name.as_ref().to_string()).await {
-                    Ok(()) => {}
-                    Err(err) => {
-                        let netns_name = netns_name.as_ref();
-                        panic!("failed to remove network namespace {netns_name}: {err}");
-                    }
-                }
-            });
-        });
-        ret.unwrap()
-    }
-}
-
-/// Fixture which creates and runs a test in the network namespace of the given name.
-fn in_scoped_netns<F: 'static + Send + RefUnwindSafe + UnwindSafe + Send + FnOnce() -> T, T>(
-    netns_name: impl 'static + Sync + UnwindSafe + RefUnwindSafe + AsRef<str>,
-) -> impl FnOnce(F) -> T
-where
-    T: Send + UnwindSafe + RefUnwindSafe,
-{
-    let netns_name_copy = netns_name.as_ref().to_string();
-    |f: F| with_scoped_netns(netns_name_copy.clone())(|| run_in_netns(netns_name_copy)(f))
-}
-
-/// Fixture which runs the supplied function with _additional_ granted capabilities.
-fn with_caps<F: UnwindSafe + FnOnce() -> T, T>(
-    caps: impl IntoIterator<Item = Capability>,
-) -> impl FnOnce(F) -> T {
-    move |f: F| {
-        let current_caps = match caps::read(None, CapSet::Effective) {
-            Ok(current_caps) => current_caps,
-            Err(err) => {
-                error!("caps error: {}", err);
-                panic!("caps error: {err}");
-            }
-        };
-        let needed_caps: Vec<_> = caps
-            .into_iter()
-            .filter(|cap| !current_caps.contains(cap))
-            .collect();
-        for cap in &needed_caps {
-            caps::raise(None, CapSet::Effective, *cap)
-                .unwrap_or_else(|err| panic!("unable to raise capability to {cap}: {err}"));
-        }
-        let ret = catch_unwind(f);
-        for cap in &needed_caps {
-            caps::drop(None, CapSet::Effective, *cap)
-                .unwrap_or_else(|err| panic!("unable to drop capability to {cap}: {err}"));
-        }
-        ret.unwrap()
-    }
-}
 
 #[test]
 #[wrap(with_caps([Capability::CAP_NET_ADMIN]))]
